Moved point and triangle to new package
