+++ lab5/Triangle.java
@@ -1,4 +1,4 @@
-package lab4;
+package lab5;
 
 public class Triangle {
 	Point A;
